Delete Tenant created page and add new Dashboard page to complete account registration flow
