+++ application/account-management/WebApp/app/dashboard/page.tsx
@@ -1,0 +1,10 @@
+export default function CreatedTenantSuccessPage() {
+  return (
+    <div className="items-center flex flex-col justify-center h-full">
+      <div className="p-8 bg-gray-800 text-white rounded-xl shadow-md text-center gap-4 flex flex-col">
+        <h1 className="text-2xl">Success!</h1>
+        <p>Your account has been created.</p>
+      </div>
+    </div>
+  );
+}
