+++ application/account-management/WebApp/translations/locale/en-US.po
@@ -14,26 +14,6 @@ msgstr ""
 "Plural-Forms: \n"
 
 #. js-lingui-explicit-id
-msgid "Missing Fields. Failed to Create Account."
-msgstr "Missing Fields. Failed to Create Account."
-
-#. js-lingui-explicit-id
-msgid "Server error: Failed to Create Account."
-msgstr "Server error: Failed to Create Account."
-
-#. js-lingui-explicit-id
-msgid "subdomain"
-msgstr "subdomain"
-
-#. js-lingui-explicit-id
-msgid "name"
-msgstr "name"
-
-#. js-lingui-explicit-id
-msgid "email"
-msgstr "email"
-
-#. js-lingui-explicit-id
 msgid "Missing Fields. Failed to login."
 msgstr "Missing Fields. Failed to login."
 
@@ -87,12 +67,6 @@ msgstr "Already have an account?"
 msgid "Can't find your code? Check your spam folder"
 msgstr "Can't find your code? Check your spam folder"
 
-msgid "Create a account"
-msgstr "Create a account"
-
-msgid "Create account!"
-msgstr "Create account!"
-
 msgid "Create your account"
 msgstr "Create your account"
 
@@ -114,9 +88,6 @@ msgstr "First name"
 msgid "Last name"
 msgstr "Last name"
 
-msgid "Name"
-msgstr "Name"
-
 msgid "Password"
 msgstr "Password"
 
@@ -137,9 +108,6 @@ msgstr "Sign up"
 
 msgid "Something went wrong, please try again"
 msgstr "Something went wrong, please try again"
-
-msgid "Subdomain"
-msgstr "Subdomain"
 
 msgid "This is the example for the<0>Acme</0>demo product, brought to you by PlatformPlatform"
 msgstr "This is the example for the<0>Acme</0>demo product, brought to you by PlatformPlatform"
